montar al app del lado del servidor, y crear las rutas con Router
