+++ package.json
@@ -20,6 +20,7 @@
     "dotenv": "^9.0.2",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "express": "^4.17.1",
+    "history": "^5.0.0",
     "md5": "^2.3.0",
     "nodemon": "^2.0.7",
     "prop-types": "^15.7.2",
@@ -27,6 +28,8 @@
     "react-dom": "^17.0.2",
     "react-hot-loader": "^4.13.0",
     "react-redux": "^7.2.4",
+    "react-router": "^5.2.0",
+    "react-router-config": "^5.1.1",
     "react-router-dom": "^5.2.0",
     "redux": "^4.1.0",
     "webpack": "^5.37.0",
