+++ src/server/server.js
@@ -21,7 +21,21 @@ if (ENV === 'development') {
 }
 
 app.get('*', (req, res) => {
-  res.send({ hello: 'express' });
+  res.send(`<!DOCTYPE html>
+  <html lang="en">
+    <head>
+      <meta charset="UTF-8" />
+      <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+      <link  rel="stylesheet" href="assets/app.css" type="text/css">
+      <title>Platzi Video</title>
+    </head>
+    <body>
+      <div id="root"></div>
+      <script src="assets/app.js" type="text/javascript"></script>
+    </body>
+  </html>
+  `);
 });
 
 app.listen(PORT, (err) => {
