+++ webpack.config.js
@@ -1,6 +1,5 @@
 const path = require('path');
 const webpack = require('webpack');
-const HtmlWebPackPlugin = require('html-webpack-plugin')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 
 module.exports = {
@@ -8,7 +7,7 @@ module.exports = {
     mode:'development',
     output: {
         path: path.resolve(__dirname, 'dist'),
-        filename: 'bundle.js',
+        filename: 'assets/app.js',
         publicPath: "/"
     },
     // extensiones que me permitirá 
@@ -62,12 +61,9 @@ module.exports = {
     },
     plugins: [
         new webpack.HotModuleReplacementPlugin(),
-        new HtmlWebPackPlugin({
-            template: './public/index.html',
-            filename:'./index.html',
-        }),
+       
         new MiniCssExtractPlugin({
-            filename:'assets/[name].css',
+            filename:'assets/app.css',
         }),
     ],
 }
